backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
openai==1.3.7
transformers==4.36.2
torch==2.1.2
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3